Test odometry in optitrack
Changed airframe xml to use optitrack.

Odometry seems to work quite ok in the cyberzoo! Final error is
about 1 meter.

Changed the coefficient scale, might need to fiddle with this a bit
more...

D term does not seem to do much for homing, keep at low value.
Filter gain might be increased for responsiveness, but is a bit
noisier at 0.50.