Revert threshold, yaw rate. Reset timeout on wrong estimates.
Reverted threshold to tighter value for better control.

Reverted yaw rate to higher values to decrease the period in which
faulty measurements occur (hopefully).

The timeout is now also reset when the velocity measurement does
not pass the sanity check. Not ideal, but at least it still
fulfills its function to verify that eVO is still running and
the USB cable is still connected. Need to find a better solution
for this in the future.