Disable file logger autostart. Voliere flight plan
Disabled automatic start of file logger. Forgot to turn the drone
of and filled the entire disk...

Simplified the percevite flightplan in percevite_voliere. The
PercEvite waypoint is hidden so it can't be moved by accident.
Removed the Flight Arena sector as it is not applicable here
or during outdoor tests. Changed PerceviteGo to PerceviteStay
so that the drone will keep following the GOAL waypoint until
manually put in standby.